feat: Erro: altura e peso estão como Byte, o que pode ser um problema pois Byte vai de -128 a 127.

+++ services/static/src/main/java/soccervs/statics/core/entities/Jogadores.java
@@ -33,10 +33,10 @@ public class Jogadores {
 	private EPePreferido pePreferido;
 	
 	@Column
-	private Byte altura;
+	private Float altura;
 	
 	@Column
-	private Byte peso;
+	private Float peso;
 	
 	@OneToMany(mappedBy = "jogadores")
 	private Set<Posicoes> posicao;
